Blueprintを利用するためにフォルダ構成修正・app_factoryに追記
